% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customHeader.R
\name{customHeader}
\alias{customHeader}
\title{customHeader}
\usage{
customHeader(app_name, top_left, bottom_left, top_right, bottom_right, left_icon, right_icon)
}
\arguments{
\item{app_name}{Character indicating App Name in the middle of the header.}

\item{top_left}{Character indicating something to put on top left. Example: Team name}

\item{top_right}{Character indicating something to put on top right. Example: Contributors to the appplication}

\item{bottom_right}{Character indicating something to put on bottom right. Example: last updated timestamp}

\item{bottom_left}{Character indicating something to put on bottom left. Example: email address}
}
\value{
Provides a standardized header for a polished look in all applications.
}
\description{
customHeader
}
\examples{
x <- "Political Candidate Favorability by U.S. State App, San Francisco Policy Insitute, Dr. Freddie Mercury and Attorney Lebron James, john.doe@email.com, 02/02/2020 02:20:20"
}
