% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customHeader.R
\name{customHeader}
\alias{customHeader}
\title{customHeader}
\usage{
customHeader(app_name, team_name, app_maker_name, email_address, last_updated)
}
\arguments{
\item{app_name}{Character indicating App Name}

\item{top_right}{Character indicating Team name}

\item{bottom_right}{Character listing contributors to the app}

\item{top_left}{Character indicating correspondant email address}

\item{bottom_left}{Character indicating the timestamp}
}
\value{
A HTML Navbar component that can be rendered through renderUI
}
\description{
customHeader
}
\examples{
x <- "Projext X,X-Men,John Doe and Bruce Wayne,john.doe@email.com, 02/02/2020 02:20:20"

}
